@@ -16,64 +16,59 @@
     "node": ">=6.5"
   },
   "scripts": {
-    "build": "cross-env NODE_ENV=production webpack --display-chunks --progress",
-    "clean:coverage": "del-cli ./coverage",
-    "clean:dist": "del-cli ./dist",
-    "lint": "run-s lint:ts",
-    "lint:ts": "tslint './src/**/*.ts'",
-    "prebuild": "run-s clean:dist lint",
-    "server:dev": "cross-env NODE_ENV=development webpack-dev-server",
-    "start": "npm run server:dev"
+    "start": "webpack-dev-server --inline --progress --port 3000",
+    "build": "rimraf dist && webpack --config config/webpack.prod.js --progress --profile --bail"
   },
   "dependencies": {
-    "@angular/common": "2.0.0",
-    "@angular/compiler": "2.0.0",
-    "@angular/core": "2.0.0",
-    "@angular/forms": "2.0.0",
-    "@angular/platform-browser": "2.0.0",
-    "@angular/platform-browser-dynamic": "2.0.0",
-    "@angular/router": "3.0.0",
-    "angularfire2": "2.0.0-beta.5",
-    "bootstrap": "4.0.0-alpha.2",
-    "bootstrap-loader": "^1.0.9",
-    "core-js": "~2.4.1",
-    "firebase": "3.3.0",
-    "jquery": "2.2.4",
-    "rxjs": "5.0.0-beta.12",
-    "tether": "1.3.2",
-    "zone.js": "~0.6.25"
+    "@angular/common": "^4.0.0",
+    "@angular/compiler": "^4.0.0",
+    "@angular/core": "^4.0.0",
+    "@angular/forms": "^4.0.0",
+    "@angular/http": "^4.0.0",
+    "@angular/platform-browser": "^4.0.0",
+    "@angular/platform-browser-dynamic": "^4.0.0",
+    "@angular/router": "^4.0.0",
+    "@angular/upgrade": "^4.0.0",
+    "angularfire2": "^4.0.0-rc.2",
+    "bootstrap": "4.0.0-alpha.6",
+    "core-js": "^2.4.1",
+    "firebase": "^4.2.0",
+    "promise-polyfill": "6.0.2",
+    "reflect-metadata": "^0.1.3",
+    "rxjs": "^5.0.1",
+    "zone.js": "^0.8.5"
   },
   "devDependencies": {
-    "@types/core-js": "~0.9.33",
-    "@types/node": "~6.0.39",
-    "@types/request": "0.0.30",
-    "autoprefixer": "~6.4.1",
-    "cross-env": "~2.0.1",
-    "css-loader": "^0.25.0",
-    "del-cli": "~0.2.0",
-    "eslint": "~3.5.0",
-    "extract-text-webpack-plugin": "2.0.0-beta.4",
-    "file-loader": "^0.9.0",
-    "firebase-tools": "3.0.7",
-    "html-webpack-plugin": "~2.22.0",
-    "minx": "r-park/minx.git",
-    "morgan": "~1.7.0",
-    "node-sass": "^3.10.0",
-    "nodemon": "~1.10.2",
-    "npm-run-all": "~3.1.0",
-    "postcss-loader": "~0.13.0",
-    "raw-loader": "~0.5.1",
-    "resolve-url-loader": "^1.6.0",
-    "sass-loader": "^4.0.2",
-    "style-loader": "^0.13.1",
-    "ts-helpers": "~1.1.1",
-    "ts-loader": "~0.8.2",
-    "tslint": "~3.15.1",
-    "typescript": "~2.0.2",
-    "typings": "~1.3.3",
-    "url-loader": "^0.5.7",
-    "webpack": "2.1.0-beta.22",
-    "webpack-dev-server": "2.1.0-beta.4",
-    "webpack-md5-hash": "~0.0.5"
+    "@types/hammerjs": "^2.0.33",
+    "@types/jasmine": "^2.5.53",
+    "@types/node": "^4.0.22-alpha",
+    "@types/request": "2.0.1",
+    "angular2-template-loader": "^0.6.2",
+    "awesome-typescript-loader": "^3.2.3",
+    "bootstrap-loader": "2.1.0",
+    "copy-webpack-plugin": "4.0.1",
+    "css-loader": "^0.28.0",
+    "exports-loader": "^0.6.4",
+    "extract-text-webpack-plugin": "2.1.0",
+    "file-loader": "0.10.1",
+    "html-loader": "0.4.5",
+    "html-webpack-plugin": "2.28.0",
+    "imports-loader": "^0.7.1",
+    "jasmine-core": "2.5.2",
+    "json-loader": "0.5.4",
+    "node-sass": "^4.5.2",
+    "postcss-loader": "^2.0.6",
+    "raw-loader": "0.5.1",
+    "resolve-url-loader": "^2.1.0",
+    "rimraf": "2.6.1",
+    "sass-loader": "^6.0.3",
+    "style-loader": "^0.18.2",
+    "ts-helpers": "1.1.2",
+    "typescript": "2.2.2",
+    "url-loader": "^0.5.9",
+    "webpack": "2.3.2",
+    "webpack-dev-server": "2.4.2",
+    "webpack-md5-hash": "0.0.5",
+    "webpack-merge": "4.1.0"
   }
 }
