@@ -1,11 +1,11 @@
 /*
  * Angular Imports
  */
-import {Component} from "@angular/core";
+import {Component} from '@angular/core';
 
 @Component({
-  selector: "my-app",
-  template: require("./app.component.html")
+  selector: 'my-app',
+  templateUrl: './app.component.html',
 })
 export class AppComponent {
 }
