@@ -21,7 +21,7 @@ import { CheckoutViewComponent } from "./checkout/checkout-view.component";
 /*
  * Routes
  */
-const routes: Routes = [
+export const routes: Routes = [
   { path: "", redirectTo: "welcome", pathMatch: "full" },
   { path: "login", component: SignInComponent },
   { path: "welcome", component: WelcomeComponent, canActivate: [AuthGuard] },
@@ -30,8 +30,3 @@ const routes: Routes = [
   { path: "cart", component: CartViewComponent, canActivate: [AuthGuard] },
   { path: "checkout", component: CheckoutViewComponent, canActivate: [AuthGuard] },
 ];
-
-/*
- * Routes Provider
- */
-export const routing = RouterModule.forRoot(routes);
