@@ -3,7 +3,8 @@
  */
 import { Injectable } from "@angular/core";
 import { CanActivate, Router } from "@angular/router";
-import { FirebaseAuthState } from "angularfire2";
+import { AngularFireAuth } from 'angularfire2/auth';
+import * as firebase from 'firebase/app';
 
 import { Observable } from "rxjs/Observable";
 import "rxjs/add/operator/do";
@@ -21,13 +22,11 @@ export class AuthGuard implements CanActivate {
     constructor(private auth: AuthService, private router: Router) { }
 
     canActivate(): Observable<boolean>|boolean {
-        return this.auth.auth$.map((authState: FirebaseAuthState) => {
-            if (authState) {
-                return true;
-            } else {
-                this.router.navigateByUrl("/login");
-                return false;
-            }
-        }).first();
+        if (this.auth.authenticated) {
+            return true;
+        } else {
+            this.router.navigateByUrl("/login");
+            return false;
+        }
     }
 }
