@@ -1,17 +1,17 @@
 /*
  * Angular Imports
  */
-import {Component, Input} from "@angular/core";
+import {Component, Input} from '@angular/core';
 
 /*
  * Components
  */
-import {Cart, CartItem, CartService} from "./cart.service";
-import {CartItemCountComponent} from "./cart-item-count.component";
+import {Cart, CartItem, CartService} from './cart.service';
+import {CartItemCountComponent} from './cart-item-count.component';
 
 @Component({
-    selector: "db-cart-view",
-    template: require("./cart-view.component.html")
+    selector: 'db-cart-view',
+    templateUrl: './cart-view.component.html'
 })
 export class CartViewComponent {
 
@@ -25,8 +25,8 @@ export class CartViewComponent {
         this.cartService.clearCart();
     }
 
-    update(value, item: CartItem) {
-        let res = value - item.count;
+    update(value: number, item: CartItem) {
+        let res: number = value - item.count;
         if (res > 0) {
             for (let i = 0; i < res; i++) {
                 this.cartService.addProduct(item.product);
