@@ -1,12 +1,12 @@
 /*
  * Angular Imports
  */
-import {Pipe, PipeTransform} from "@angular/core";
+import {Pipe, PipeTransform} from '@angular/core';
 
 /*
  * Components Imports
  */
-import {Category, CategoryService} from "./category.service";
+import {Category, CategoryService} from './category.service';
 
 /*
  * Return category title of the value
@@ -14,16 +14,16 @@ import {Category, CategoryService} from "./category.service";
  *   value | categoryTitle
  * Example:
  *   {{ categoryId |  categoryTitle }}
- *   presume categoryId="1"
- *   result formats to "Bread & Bakery"
+ *   presume categoryId='1'
+ *   result formats to 'Bread & Bakery'
 */
-@Pipe({ name: "categoryTitle" })
+@Pipe({ name: 'categoryTitle' })
 export class CategoryTitlePipe implements PipeTransform {
 
     constructor(private categoryService: CategoryService) { }
 
     transform(value: string): string {
         let category: Category = this.categoryService.getCategory(value);
-        return category ? category.title : "";
+        return category ? category.title : '';
     }
 }
