@@ -1,11 +1,11 @@
 /*
  * Components
  */
-import {Component} from "@angular/core";
+import {Component} from '@angular/core';
 
 @Component({
-  selector: "db-footer",
-  template: require("./footer.component.html")
+  selector: 'db-footer',
+  templateUrl: './footer.component.html'
 })
 export class FooterComponent {
 }
