@@ -1,16 +1,16 @@
 /*
  * Angular Imports
  */
-import {Component, Input, Output, EventEmitter} from "@angular/core";
+import {Component, Input, Output, EventEmitter} from '@angular/core';
 
 /*
  * Components
  */
-import {Product} from "./product.service";
+import {Product} from './product.service';
 
 @Component({
-    selector: "db-product-card",
-    template: require("./product-card.component.html")
+    selector: 'db-product-card',
+    templateUrl: './product-card.component.html'
 })
 export class ProductCardComponent {
     @Input() products: Product[];
@@ -18,8 +18,8 @@ export class ProductCardComponent {
 
     setClasses(product: Product) {
         return {
-            "card-danger": product.isSpecial,
-            "card-inverse": product.isSpecial
+            'card-danger': product.isSpecial,
+            'card-inverse': product.isSpecial
         };
     }
 
