@@ -1,10 +1,10 @@
 /*
  * Angular Imports
  */
-import {Component} from "@angular/core";
+import {Component} from '@angular/core';
 
 @Component({
-  selector: "db-products",
-  template: require("./product-list.component.html")
+  selector: 'db-products',
+  templateUrl: './product-list.component.html'
 })
 export class ProductListComponent {}
