@@ -1,10 +1,10 @@
 /*
  * Angular Imports
  */
-import {Directive, Input} from "@angular/core";
-import {TemplateRef, ViewContainerRef} from "@angular/core";
+import {Directive, Input} from '@angular/core';
+import {TemplateRef, ViewContainerRef} from '@angular/core';
 
-@Directive({ selector: "[dbIf]" })
+@Directive({ selector: '[dbIf]' })
 export class IfDirective {
     constructor(
         private templateRef: TemplateRef<any>,
