@@ -1,9 +1,9 @@
-import {FormControl} from "@angular/forms";
-import {Directive, forwardRef} from "@angular/core";
-import {NG_VALIDATORS} from "@angular/forms";
+import {FormControl} from '@angular/forms';
+import {Directive,forwardRef} from '@angular/core';
+import {NG_VALIDATORS} from '@angular/forms';
 
 export function validateZip(c: FormControl) {
-  let ZIP_REGEXP: RegExp = new RegExp("[A-Za-z]{5}");
+  let ZIP_REGEXP:RegExp = new RegExp('[A-Za-z]{5}');
 
   return ZIP_REGEXP.test(c.value) ? null : {
     validateZip: {
@@ -13,7 +13,7 @@ export function validateZip(c: FormControl) {
 }
 
 @Directive({
-  selector: "[validateZip][ngModel],[validateZip][formControl]",
+  selector: '[validateZip][ngModel],[validateZip][formControl]',
   providers: [
     { provide: NG_VALIDATORS, useExisting: forwardRef(() => ZipValidator), multi: true }
   ]
